pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.2"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "validators>=0.34.0",
]

[dependency-groups]
dev = [
    "flake8>=7.1.2",
    "isort>=6.0.1",
]
